remove hardcoded redirect url variable in productorder model and replace with setting variable. refactor payment api endpoint url

+++ src/hkm/hkm/printmotor_client.py
@@ -49,7 +49,7 @@ class PrintmotorClient(object):
                 } 
             ],
             'endUserPrice' : { 
-              'priceValue' : int(order.total_price), # int ??,
+              'priceValue' : float(order.total_price),
               'currencyIso4217' : "EUR",
             }
           }
